Fix some more stuff with subscriptions.

+++ src/main/java/com/bubble/db/subscription/SubsRepository.java
@@ -2,8 +2,8 @@ package com.bubble.db.subscription;
 
 public interface SubsRepository {
 	
-	public void createSubscription(int topic, String user, boolean subscribed);
+	public void createSubscription(long topic, String user, boolean subscribed);
 
-	public boolean isSubscribed(String user, int topic);
+	public boolean isSubscribed(String user, long topic);
 
 }
